# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe API Docs",
        "description": "API documentation for the Recipe Application",
        "contact": {
            "name": "Recipe App Support",
            "email": "support@recipeapp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9020",
            "description": "Version 1 of the Recipe API"
        },
        {
            "url": "http://localhost:9020/"
        }
    ],
    "paths": {
...
        "/api/v1/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "List authenticated user's recipes",
                "description": "Retrieve all recipes created by the authenticated user. Pagination is not included in this endpoint.",
                "operationId": "2450a29256346aa792a7761b0e0fe57d",
                "responses": {
                    "200": {
                        "description": "List of user's recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Spaghetti Carbonara"
                                            },
                                            "cuisine_type": {
                                                "type": "string",
                                                "example": "Italian"
                                            },
                                            "ingredients": {
                                                "type": "string",
                                                "example": "200g pasta\\n4 eggs\\n100g pancetta"
                                            },
                                            "steps": {
                                                "type": "string",
                                                "example": "1. Boil pasta\\n2. Fry pancetta\\n3. Mix with eggs"
                                            },
                                            "picture": {
                                                "type": "string",
                                                "example": "recipes/abc123.jpg",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-04-05T10:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-04-05T12:30:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "operationId": "d03f498cebb831ef1d1112b1be6bc4c9",
                "requestBody": {
                    "description": "Recipe data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cuisine_type",
                                    "ingredients",
                                    "steps"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Recipe name",
                                        "type": "string",
                                        "example": "Spaghetti Carbonara"
                                    },
                                    "cuisine_type": {
                                        "description": "Cuisine type",
                                        "type": "string",
                                        "example": "Italian"
                                    },
                                    "ingredients": {
                                        "description": "Ingredients (one per line)",
                                        "type": "string",
                                        "example": "200g pasta\\n4 eggs\\n100g pancetta"
                                    },
                                    "steps": {
                                        "description": "Steps (one per line)",
                                        "type": "string",
                                        "example": "1. Boil pasta\\n2. Fry pancetta\\n3. Mix with eggs"
                                    },
                                    "picture": {
                                        "description": "Recipe image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Spaghetti Carbonara"
                                        },
                                        "cuisine_type": {
                                            "type": "string",
                                            "example": "Italian"
                                        },
                                        "ingredients": {
                                            "type": "string",
                                            "example": "200g pasta\\n4 eggs\\n100g pancetta"
                                        },
                                        "steps": {
                                            "type": "string",
                                            "example": "1. Boil pasta\\n2. Fry pancetta\\n3. Mix with eggs"
                                        },
                                        "picture": {
                                            "type": "string",
                                            "example": "recipes/abc123.jpg",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum for API authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}